requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
anthropic
itsdangerous
jinja2
python-multipart
python-dotenv
